# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "忆述光华 API",
    "description": "个人记忆整理与回顾后端。功能：认证（微信/手机号/JWT）、内容入库（照片/文字/语音）、四层事件聚合、描述性搜索（RAG）、同步（LWW）、AI 分类/画像/护栏。契约对齐《忆述光华_Sprint1规划.md》S1-02。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/wechat/find": {
      "post": {
        "tags": [
          "wechat"
        ],
        "summary": "Wechat Find",
        "description": "微信\"找\"：沙箱/客户端直接调用（10s/3s 门禁在沙箱可测 WX-007）",
        "operationId": "wechat_find_api_v1_wechat_find_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatFindRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/wechat/callback": {
      "get": {
        "tags": [
          "wechat"
        ],
        "summary": "Wechat Callback Verify",
        "description": "URL 验证：解密 echostr 原样返回（企微 GET 请求）",
        "operationId": "wechat_callback_verify_api_v1_wechat_callback_get",
        "parameters": [
          {
            "name": "msg_signature",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Msg Signature"
            }
          },
          {
            "name": "timestamp",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timestamp"
            }
          },
          {
            "name": "nonce",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Nonce"
            }
          },
          {
            "name": "echostr",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Echostr"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "wechat"
        ],
        "summary": "Wechat Callback",
        "description": "收消息：验签→解密→幂等入库→返回 success（企微要求响应\"success\"或空串）",
        "operationId": "wechat_callback_api_v1_wechat_callback_post",
        "parameters": [
          {
            "name": "msg_signature",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Msg Signature"
            }
          },
          {
            "name": "timestamp",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timestamp"
            }
          },
          {
            "name": "nonce",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Nonce"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wechat/delete": {
      "post": {
        "tags": [
          "wechat"
        ],
        "summary": "Wechat Delete",
        "description": "微信端软删本条（F6：只删本条；审查 CRITICAL 修复：补鉴权 + 归属校验）",
        "operationId": "wechat_delete_api_v1_wechat_delete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "msg_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Msg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse",
        "description": "统一响应体：{code, message, data, request_id}"
      },
...
      "ApiResponse_dict_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[dict]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WechatFindRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Query"
          },
          "limit": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "WechatFindRequest",
        "description": "微信\"找\"请求（S4-02：消息解析 → RAG 搜索）"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}